requirements.txt
functions-framework==3.*
google-cloud-bigquery>=2.20.0
requests>=2.31.0
